+++ assets/js/modules/session/operations.js
@@ -18,18 +18,29 @@ const changeUsername = username => {
   }
 }
 
-const updateAccount = (data) => {
+const changeEmail = data => {
   return ( dispatch ) => {
     dispatch(start_request())
 
-    privateChannel.push("update_account", data)
+    privateChannel.push("change_email", data)
+    .receive('ok', res => dispatch(Flash.get(res)))
+    .receive('error', res => dispatch(Flash.get(res)))
+  }
+}
+
+const changePassword = data => {
+  return ( dispatch ) => {
+    dispatch(start_request())
+
+    privateChannel.push("change_password", data)
     .receive('ok', res => dispatch(Flash.get(res)))
     .receive('error', res => dispatch(Flash.get(res)))
   }
 }
 
 export default {
+  changeEmail,
+  changePassword,
   changeUsername,
-  login,
-  updateAccount
+  login
 };