+++ lib/crowd_crush_web/channels/private_channel.ex
@@ -15,7 +15,7 @@ defmodule CrowdCrushWeb.PrivateChannel do
 
   def handle_in("change_username", username, socket) do
     case Accounts.update_user(socket.assigns.user_id, %{ username: username }) do
-      {:ok, user } ->
+      {:ok, _user } ->
         return_success socket,
           "Your username was successfully changed to: #{username}",
           username: username
@@ -24,21 +24,40 @@ defmodule CrowdCrushWeb.PrivateChannel do
     end
   end
 
-  def handle_in("update_account", %{"password" => password} = params, socket) do
+  def handle_in("change_email", %{"email" => email, "password" => password}, socket) do
     user = Accounts.get_user_with_credential(socket.assigns.user_id)
-    case Pbkdf2.verify_pass(password, user.credential.password_hash) do
-      true ->
-        changes = Map.put(params, "password", params["new_password"])
 
-        case Accounts.update_credential(user.credential, changes) do
+    if email == user.credential.email do
+      {:reply, {:error, %{ info: "Your new email matches your current email address. No change was conducted." }}, socket}
+    else
+      case Pbkdf2.verify_pass(password, user.credential.password_hash) do
+        true ->
+          case Accounts.update_credential(user.credential, %{ email: email }) do
+            {:ok, _credential } ->
+              {:reply, {:ok, %{success: "You have changed your email." }}, socket}
+            {:error, changeset} ->
+              Logger.warn "#{inspect changeset}"
+              {:reply, {:error, %{ error: "Could not update email!" }}, socket}
+          end
+        false ->
+          {:reply, {:error, %{ error: "Wrong password!" }}, socket}
+      end
+    end
+  end
+
+  def handle_in("change_password", %{"new_password" => password, "password" => pass}, socket) do
+    user = Accounts.get_user_with_credential(socket.assigns.user_id)
+    case Pbkdf2.verify_pass(pass, user.credential.password_hash) do
+      true ->
+        case Accounts.update_credential(user.credential, %{ password: password }) do
           {:ok, _credential } ->
-            {:reply, {:ok, %{success: "You have changed your email/password." }}, socket}
+            {:reply, {:ok, %{success: "You have changed your password." }}, socket}
           _ ->
-            {:reply, {:error, %{ error: "Could not update email/password!" }}, socket}
+            {:reply, {:error, %{ error: "Could not update password!" }}, socket}
         end
       false ->
         {:reply, {:error, %{ error: "Wrong password!" }}, socket}
-      end
+    end
   end
 
   def handle_in("add_video", params, socket) do